+++ bootique-tapestry55/src/main/java/io/bootique/tapestry/v55/TapestryModuleExtender.java
@@ -30,6 +30,10 @@ import io.bootique.tapestry.v55.annotation.Symbols;
 import io.bootique.tapestry.v55.annotation.TapestryModuleBinding;
 import org.apache.tapestry5.services.LibraryMapping;
 
+/**
+ * @deprecated in favor of 5.8 (or later) modules
+ */
+@Deprecated(since = "3.0", forRemoval = true)
 public class TapestryModuleExtender extends ModuleExtender<TapestryModuleExtender> {
 
     private MapBuilder<String, String> symbols;