+++ bootique-tapestry55/src/main/java/io/bootique/tapestry/v55/di/InjectorServiceDef.java
@@ -28,6 +28,10 @@ import org.apache.tapestry5.ioc.def.ServiceDef;
 import java.util.Collections;
 import java.util.Set;
 
+/**
+ * @deprecated in favor of 5.8 (or later) modules
+ */
+@Deprecated(since = "3.0", forRemoval = true)
 public class InjectorServiceDef implements ServiceDef {
 
     private Injector injector;