+++ bootique-tapestry55/src/main/java/io/bootique/tapestry/v55/env/TapestryEnvironment.java
@@ -27,7 +27,10 @@ import java.util.Optional;
  * Provides access to Tapestry {@link Registry} in a context of a given web application. Allows Bootique DI services to
  * lookup Tapestry-specific services that otherwise can't be injected. Will throw an exception if the Tapestry subsystem
  * is not yet started, and its servlet filters are not yet initialized.
+ *
+ * @deprecated in favor of 5.8 (or later) modules
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public interface TapestryEnvironment {
 
     Optional<Registry> getRegistry();