+++ bootique-tapestry55/src/main/java/io/bootique/tapestry/v55/filter/BQTapestryFilter.java
@@ -45,7 +45,10 @@ import java.io.IOException;
 
 /**
  * A servlet filter that starts Tapestry environment. Based on Tapestry {@link org.apache.tapestry5.TapestryFilter}.
+ *
+ * @deprecated in favor of 5.8 (or later) modules
  */
+@Deprecated(since = "3.0", forRemoval = true)
 // we couldn't subclass TapestryFilter as the interesting parts (init() method) are final and too coarse grained.
 // also our own filter allows us to switch all the configuration to Bootique style.
 public class BQTapestryFilter implements Filter {